# refactor: rephrase the register frame-condition precondition as non-membership in a list of modified registers